pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "e2neutrino"
version = "0.1.0"
description = "Convert Enigma2 settings into Neutrino (Zapit) profiles."
readme = "README.en.md"
license = "MIT"
authors = [
  { name = "Neutrino Settings Automation", email = "team@example.com" }
]
requires-python = ">=3.10"
dependencies = [
  "click==8.3.0",
  "pyyaml==6.0.3",
  "requests==2.32.5",
  "jsonschema==4.23.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.2",
  "ruff==0.14.0",
  "mypy==1.18.2",
  "build==1.3.0",
  "types-PyYAML==6.0.12.20250915",
]

[project.scripts]
e2neutrino = "e2neutrino.__main__:cli"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["e2neutrino*"]
exclude = ["tests", "samples", "examples"]

[tool.setuptools.package-data]
"e2neutrino.schemas" = ["*.json"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
